+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Create an `assets/js/pages` directory in the Igniter install task and fix the documentation ([#57](https://github.com/inertiajs/inertia-phoenix/pull/57)).
 
+### Fixed
+
+- Properly camelize keys in `deferredProps` metadata when `camelize_props` is enabled.
+
 ## 2.5.1
 
 ### Fixed